Test works now.  Needed to pause to let PWM do its thing

+++ Software/DumptruckUltra-MCU/hw/motors/Motor.hpp
@@ -23,7 +23,7 @@ public:
     void disable();
     void setPower(float power);
 
-    static constexpr uint32_t MOTOR_MAX_SPEED_ABS{100};
+    static constexpr float MOTOR_MAX_SPEED_ABS{100};
 
 private:
     cyhal_pwm_t forwardPWMHandle;
